Fix bad commit - (Added Images)
Fixes the bad commit that had stashed changes as part of the Class. Prev. Notes:

Images are held as Final Strings in the Images class, and are printed out when needed using a function in the Helper Class.

There are still temporary images, and the rest need to be added.